no-op refactoring change - split console.py into separate modules for each capability as it's long and hard to navigate otherwise; the imports in console.py ensure the same symbols are available to protect compatibility
